tester: Set CURLOPT_TIMEOUT to a reasonable value
If we don't want to grow the list faster than we can empty it, the
timeout needs to be a few ms shorter than the time it takes us to fill
back in the list.  Sadly, this is probably still a broken idea but it is
better than nothing, for now.

Ideally we'd see that an easy_handle failed and then just dump the
entire list, as those are going to fail as well real soon now.  In this
way, we'd also rate limit if the host on the other end is quite slow for
some reason but at the tradeoff of dropping data.